Duplicate removal and test fixes
