+++ src/lib/util/getOSItem.ts
@@ -1,8 +1,6 @@
 import { Items } from 'oldschooljs';
 import { Item } from 'oldschooljs/dist/meta/types';
 
-import { stringMatches } from '../util';
-
 const cache = new Map();
 
 function cleanItemName(itemName: string) {
@@ -26,18 +24,4 @@ export default function getOSItem(itemName: string | number): Item {
 	if (!osItem) throw "That item doesn't exist.";
 	cache.set(itemName, osItem);
 	return osItem;
-}
-
-export function getOSItems(itemName: string | number): Item[] {
-	let osItems = [];
-
-	let identifier: string | number | undefined = '';
-	if (typeof itemName === 'number') {
-		identifier = itemName;
-		osItems = Items.filter(item => item.id === identifier).array() as Item[];
-	} else {
-		osItems = Items.filter(item => stringMatches(item.name, itemName)).array() as Item[];
-	}
-	if (osItems.length === 0) throw "That item doesn't exist.";
-	return osItems;
 }
